requirements.txt
python-telegram-bot==21.11
httpx==0.28.1
gspread==6.1.4
